Update tests cases to conform to new main core changes
